+++ src/scene_console.h
@@ -7,7 +7,7 @@
 class ConsoleScene : public Scene {
 public:
 	void Startup(ClientInfo* i) override;
-	void Update(double dt) override;
+	void Update(float dt) override;
 	void Render() override;
 	bool Event(SDL_Event *ev) override;
